requirements.txt
PySimpleGUI
Pillow
mutagen
